Fix bug: incorrect substitutions for refs to undefined macro args or \@
References to undefined macro args (\1, ...) or \@ are wrongly
substituted by the char following '\' in "" and in args of macro calls
and by the empty string in {} and "{}".

Examples:

1) In a macro call:

SECTION "HOME", HOME[0]

m: MACRO
        db \1
ENDM

        m \2
        m \@

\2 and \@ are not defined.

Object code:
	02 01
	the access to \2 has been substituted by the number 2
	and the access to \@ has been substituted by @, the PC counter.

2) In a string:

	db "\2\@"
will produce:
	32 40

3) In {}:
	db {\1}
and
	db "{\@}"
Will generate an error about an unknown symbol ''.

Solution:

The current behaviour of the assembler when it encounters an access to
an undefined macro arg in other cases (e.g. db \1) is to generate an
error and this is the solution I have adopted.

Remark:
  Error messages could be improved if we found a way to detect if we
  are in the main program or in a macro expansion as macro args are not
  supposed to be available outside of a macro definition.