Add option to mark the ROM bank number in each one
Added option -b for rgblink, which will reserve the last byte of each
ROMX bank (not ROM0), and for rgbfix, which will write the ROMX
number on said byte. It only works with ROMs with more than 2 ROM
banks and less or equal than 256 banks (because it would overflow).

This is useful for some systems of handling ROM banks.