Introduce safe libc functions
The functions available in libc return error codes that can be used
to stop the program in case of error, but checking for error at every
call makes the code unreadable in some cases.

The new functions introduced with this patch can be used to easily do
error checking at every call of a function that handles I/O or dynamic
memory.

Signed-off-by: Antonio Niño Díaz <antonio_nd@outlook.com>