Allow multiple identical exported numeric constants
This does not allow identical exported label constants, which *is*
possible using a WRAM SECTION UNION. That can be addressed when or
if it's requested for a real use case.

Symbols only store one source filename + line number, so this
arbitrarily keeps the last read symbol as the reported one.